+++ src/main/java/roomescape/reservation/repository/dto/PopularThemeQueryResult.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.repository.dto;
 
 public record PopularThemeQueryResult(Long id, String name, String description, String thumbnailUrl) {
 }
