+++ src/main/java/roomescape/reservation/service/dto/PopularThemesResult.java
@@ -1,6 +1,6 @@
 package roomescape.reservation.service.dto;
 
-import roomescape.reservation.repository.PopularThemeQueryResult;
+import roomescape.reservation.repository.dto.PopularThemeQueryResult;
 
 import java.util.List;
