+++ src/main/java/roomescape/theme/service/dto/AvailableTimesResult.java
@@ -1,7 +1,7 @@
 package roomescape.theme.service.dto;
 
 import java.util.List;
-import roomescape.time.repository.AvailableTimeQueryResult;
+import roomescape.time.repository.dto.AvailableTimeQueryResult;
 
 public record AvailableTimesResult(List<AvailableTimeQueryResult> availableTimeQueryResults) {
 }
