+++ src/main/java/roomescape/time/repository/JdbcReservationTimeRepository.java
@@ -12,6 +12,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.time.domain.ReservationTime;
+import roomescape.time.repository.dto.AvailableTimeQueryResult;
 
 @Repository
 public class JdbcReservationTimeRepository implements ReservationTimeRepository {
