+++ src/main/java/roomescape/time/repository/dto/AvailableTimeQueryResult.java
@@ -1,4 +1,4 @@
-package roomescape.time.repository;
+package roomescape.time.repository.dto;
 
 import java.time.LocalTime;
