+++ src/main/java/roomescape/time/service/dto/ReservationTimeCommand.java
@@ -1,4 +1,4 @@
-package roomescape.time.service;
+package roomescape.time.service.dto;
 
 import java.time.LocalTime;
