+++ src/test/java 2/roomescape/RoomescapeApplicationTest.java
@@ -1,0 +1,13 @@
+package roomescape;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RoomescapeApplicationTest {
+
+        @Test
+        void contextLoads() {
+        }
+
+}
